# perf: minor adjustments