Minor adjustments to finally reach the 1M/s mark 🎉

```sh
$ time cargo r --release --bin generator -- -n 5000000 -j 2 | time cargo r --release
       Total 5 000 000 order(s) in 5.00s
     Average 1 000 051.81 orders/s

 Orderbook info
    Spread
     Ask 7 363
     Bid 4 889
    Length
     Ask 548 050
     Bid 541 117

        5.50 real         9.47 user         0.32 sys
cargo r --release --bin generator -- -n 5000000 -j 2  4.49s user 0.80s system 99% cpu 5.325 total
time cargo r --release  9.47s user 0.33s system 177% cpu 5.507 total